requirements.txt
gradio==4.44.0
openai-whisper
pydub
nltk
ffmpeg-python
